Support the add-ratings version.
