+++ market/add-ratings/Makefile
@@ -1,0 +1,28 @@
+all: checkVirtual prep ratingsService start
+
+checkVirtual:
+	@if [ -z "$$VIRTUAL_ENV" ]; then \
+		echo "Please activate a virtualenv to continue (cf https://virtualenv.readthedocs.org)" >&2 ;\
+		exit 1; \
+	fi
+
+prep:
+	pip install -r requirements.txt
+
+ratingsService:
+	quark install --python ../ratings/ratings.q
+
+start:
+	python market.py
+
+startRatings:
+	python ../ratings/ratingsService.py 8001
+
+test tests:
+	@echo "Make sure the Market is running, then Y to continue or N to quit:"
+	@read junk ;\
+	case $$junk in \
+	  y|Y)  nosetests ;\
+            ;; \
+      *)    ;; \
+    esac
