+++ market/add-ratings/requirements.txt
@@ -1,0 +1,4 @@
+datawire-quarkdev==0.4.13
+Flask==0.10.1
+nose==1.3.7
+requests==2.9.1
